fix: batch 2 review — async safety, WS cleanup, iteration
- Add run_sync() helper to offload all blocking SDK calls
- Wrap contract/market_data/WS subscribe calls in run_in_executor
- Snapshot-copy subscription sets before iterating in broadcast
- Return orphaned keys from disconnect(), unsubscribe on WS close
- Fix contract iteration to flat pattern (not nested exchange loop)
- Validate WS message fields before processing
- Fix SnapshotData buy_volume/sell_volume to float

+++ src/shioaji_server/models.py
@@ -80,9 +80,9 @@ class SnapshotData(BaseModel):
     volume: int
     total_volume: int
     buy_price: float
-    buy_volume: int
+    buy_volume: float
     sell_price: float
-    sell_volume: int
+    sell_volume: float
     change_price: float
     change_rate: float
     ts: int
